refactor(api): share createTestUser test helper
Extract the user-row insert into `createTestUser` in `test-utils.ts`
so both `seedUserWithFixture` (for the enrollment fixture path) and
the new `enrollment.test.ts` (for the constraint-violation path)
stop hand-rolling the same schema.user insert.
